重构了参数类，重构目的是 一个参数就可以适配常量模式和绑定模式 这样的话，在工具层的核心算法，只需要对参数进行使用，完全不用关系，使用的 参数是通过绑定表达式解析过来的，还是通过常量模式直接设置的

+++ tools/SunEyeVision.Tool.Threshold/ThresholdResults.cs
@@ -20,12 +20,12 @@ namespace SunEyeVision.Tool.Threshold
         /// <summary>
         /// 实际使用的阈值
         /// </summary>
-        public double ThresholdUsed { get; set; }
+        public int ThresholdUsed { get; set; }
 
         /// <summary>
         /// 最大值
         /// </summary>
-        public double MaxValueUsed { get; set; }
+        public int MaxValueUsed { get; set; }
 
         /// <summary>
         /// 使用的阈值类型
